# openapi/ema.v1.s16a.json
{
...
  "info": {
    "title": "Ema local API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/clients": {
      "get": {
        "operationId": "handler_clients_get",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": [
                  {
                    "anaf_refreshed_at": "2026-01-01T00:00:00Z",
                    "caen": "0000",
                    "contacts": [
                      {
                        "id": "contact-exemplu",
                        "name": "Persoană exemplu",
                        "role": "manager energetic"
                      }
                    ],
                    "cui": "00000000",
                    "id": "client-exemplu",
                    "name": "Societate Exemplu SRL",
                    "sites": [
                      {
                        "address": "Adresă sintetică",
                        "id": "site-exemplu",
                        "name": "Sediu exemplu"
                      }
                    ]
                  }
                ],
                "schema": {
                  "items": {
                    "properties": {
                      "anaf_refreshed_at": {
                        "type": "string"
                      },
                      "caen": {
                        "type": "string"
                      },
                      "contacts": {
                        "items": {
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            },
                            "role": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "id",
                            "name",
                            "role"
                          ],
                          "type": "object"
                        },
                        "type": "array"
                      },
                      "cui": {
                        "type": "string"
                      },
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "sites": {
                        "items": {
                          "properties": {
                            "address": {
                              "type": "string"
                            },
                            "id": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "id",
                            "name",
                            "address"
                          ],
                          "type": "object"
                        },
                        "type": "array"
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "cui",
                      "caen",
                      "anaf_refreshed_at",
                      "sites",
                      "contacts"
                    ],
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "501": {
            "description": "Contract only; use case pending"
          }
        },
        "summary": "Handler",
        "tags": [
          "provisional"
        ],
        "x-provisional": true
      },
      "post": {
        "operationId": "handler_clients_post",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "cui": "00000000",
                "name": "Societate Exemplu SRL"
              },
              "schema": {
                "properties": {
                  "cui": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "cui"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "anaf_refreshed_at": "2026-01-01T00:00:00Z",
                  "caen": "0000",
                  "contacts": [
                    {
                      "id": "contact-exemplu",
                      "name": "Persoană exemplu",
                      "role": "manager energetic"
                    }
                  ],
                  "cui": "00000000",
                  "id": "client-exemplu",
                  "name": "Societate Exemplu SRL",
                  "sites": [
                    {
                      "address": "Adresă sintetică",
                      "id": "site-exemplu",
                      "name": "Sediu exemplu"
                    }
                  ]
                },
                "schema": {
                  "properties": {
                    "anaf_refreshed_at": {
                      "type": "string"
                    },
                    "caen": {
                      "type": "string"
                    },
                    "contacts": {
                      "items": {
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "role": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "role"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "cui": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "sites": {
                      "items": {
                        "properties": {
                          "address": {
                            "type": "string"
                          },
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "address"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "cui",
                    "caen",
                    "anaf_refreshed_at",
                    "sites",
                    "contacts"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "501": {
            "description": "Contract only; use case pending"
          }
        },
        "summary": "Handler",
        "tags": [
          "provisional"
        ],
        "x-provisional": true
      }
    },
    "/clients/{client_id}": {
      "get": {
        "operationId": "handler_clients__client_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "anaf_refreshed_at": "2026-01-01T00:00:00Z",
                  "caen": "0000",
                  "contacts": [
                    {
                      "id": "contact-exemplu",
                      "name": "Persoană exemplu",
                      "role": "manager energetic"
                    }
                  ],
                  "cui": "00000000",
                  "id": "client-exemplu",
                  "name": "Societate Exemplu SRL",
                  "sites": [
                    {
                      "address": "Adresă sintetică",
                      "id": "site-exemplu",
                      "name": "Sediu exemplu"
                    }
                  ]
                },
                "schema": {
                  "properties": {
                    "anaf_refreshed_at": {
                      "type": "string"
                    },
                    "caen": {
                      "type": "string"
                    },
                    "contacts": {
                      "items": {
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "role": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "role"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "cui": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "sites": {
                      "items": {
                        "properties": {
                          "address": {
                            "type": "string"
                          },
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "address"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "cui",
                    "caen",
                    "anaf_refreshed_at",
                    "sites",
                    "contacts"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "501": {
            "description": "Contract only; use case pending"
          }
        },
        "summary": "Handler",
        "tags": [
          "provisional"
        ],
        "x-provisional": true
      },
      "patch": {
        "operationId": "handler_clients__client_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "caen": "0000",
                "name": "Societate Exemplu SRL"
              },
              "schema": {
                "properties": {
                  "caen": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "caen"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "anaf_refreshed_at": "2026-01-01T00:00:00Z",
                  "caen": "0000",
                  "contacts": [
                    {
                      "id": "contact-exemplu",
                      "name": "Persoană exemplu",
                      "role": "manager energetic"
                    }
                  ],
                  "cui": "00000000",
                  "id": "client-exemplu",
                  "name": "Societate Exemplu SRL",
                  "sites": [
                    {
                      "address": "Adresă sintetică",
                      "id": "site-exemplu",
                      "name": "Sediu exemplu"
                    }
                  ]
                },
                "schema": {
                  "properties": {
                    "anaf_refreshed_at": {
                      "type": "string"
                    },
                    "caen": {
                      "type": "string"
                    },
                    "contacts": {
                      "items": {
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "role": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "role"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "cui": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "sites": {
                      "items": {
                        "properties": {
                          "address": {
                            "type": "string"
                          },
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "address"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "cui",
                    "caen",
                    "anaf_refreshed_at",
                    "sites",
                    "contacts"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "501": {
            "description": "Contract only; use case pending"
          }
        },
        "summary": "Handler",
        "tags": [
          "provisional"
        ],
        "x-provisional": true
      }
    },
    "/clients/{client_id}/anaf/refresh": {
      "post": {
        "operationId": "handler_clients__client_id__anaf_refresh_post",
        "parameters": [
          {
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {},
              "schema": {
                "properties": {},
                "required": [],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "client_id": "client-exemplu",
                  "retrieved_at": "2026-01-01T00:00:00Z",
                  "source_url": "https://example.invalid/anaf",
                  "status": "complete"
                },
                "schema": {
                  "properties": {
                    "client_id": {
                      "type": "string"
                    },
                    "retrieved_at": {
                      "type": "string"
                    },
                    "source_url": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "client_id",
                    "status",
                    "retrieved_at",
                    "source_url"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "501": {
            "description": "Contract only; use case pending"
          }
        },
        "summary": "Handler",
        "tags": [
          "provisional"
        ],
        "x-provisional": true
      }
    },
    "/clients/{client_id}/contacts": {
      "get": {
        "operationId": "handler_clients__client_id__contacts_get",
        "parameters": [
          {
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": "contact-exemplu",
                    "name": "Persoană exemplu",
                    "role": "manager energetic"
                  }
                ],
                "schema": {
                  "items": {
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "role": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "role"
                    ],
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "501": {
            "description": "Contract only; use case pending"
          }
        },
        "summary": "Handler",
        "tags": [
          "provisional"
        ],
        "x-provisional": true
      }
    },
    "/clients/{client_id}/files": {
      "post": {
        "operationId": "handler_clients__client_id__files_post",
        "parameters": [
          {
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "intake": "reading",
                  "name": "exemplu.pdf",
                  "sha": "synthetic-sha"
                },
                "schema": {
                  "properties": {
                    "intake": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "sha": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "sha",
                    "name",
                    "intake"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "501": {
            "description": "Contract only; use case pending"
          }
        },
        "summary": "Handler",
        "tags": [
          "provisional"
        ],
        "x-provisional": true
      }
    },
    "/clients/{client_id}/files/{file_sha}/versions": {
      "get": {
        "operationId": "handler_clients__client_id__files__file_sha__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "file_sha",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": [
                  {
                    "name": "exemplu.pdf",
                    "sha": "synthetic-sha",
                    "size_bytes": 4096,
                    "version": 1
                  }
                ],
                "schema": {
                  "items": {
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "sha": {
                        "type": "string"
                      },
                      "size_bytes": {
                        "type": "integer"
                      },
                      "version": {
                        "type": "integer"
                      }
                    },
                    "required": [
                      "version",
                      "sha",
                      "name",
                      "size_bytes"
                    ],
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "501": {
            "description": "Contract only; use case pending"
          }
        },
        "summary": "Handler",
        "tags": [
          "provisional"
        ],
        "x-provisional": true
      }
    },
    "/clients/{client_id}/sites": {
      "get": {
        "operationId": "handler_clients__client_id__sites_get",
        "parameters": [
          {
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": [
                  {
                    "address": "Adresă sintetică",
                    "id": "site-exemplu",
                    "name": "Sediu exemplu"
                  }
                ],
                "schema": {
                  "items": {
                    "properties": {
                      "address": {
                        "type": "string"
                      },
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "address"
                    ],
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "501": {
            "description": "Contract only; use case pending"
          }
        },
        "summary": "Handler",
        "tags": [
          "provisional"
        ],
        "x-provisional": true
      }
    },
...
  }
}